# DB/app/ilv-server-all/rest-api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Defintion of ILV resources",
    "version": "0.0.1",
    "title": "ILV",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "albert.lacambra-basil@deutschebahn.com"
    }
  },
  "host": "localhost:1234",
  "basePath": "/app/resources/v0/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/offer": {
      "post": {
        "summary": "Creates a new offer",
        "description": "",
        "operationId": "createOffer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "offer contents",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InsertUpdateOffer"
            }
          },
          {
            "in": "query",
            "name": "directSubmit",
            "description": "the created offer will be directed submitted in case its valid",
            "required": false,
            "default": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "201": {
            "description": "offer created",
            "schema": {
              "$ref": "#/definitions/ObjectId"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "405": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/ApiErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ApiErrorResponse"
            }
          }
        },
        "security": [
          {
            "ilv_auth": []
          }
        ]
      }
    },
    "/offer/{id}": {
      "put": {
        "summary": "Update an existing offer",
        "description": "",
        "operationId": "updateOffer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Offer object that needs to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "demand contents",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InsertUpdateOffer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Offer updated"
          },
          "400": {
            "description": "Invalid ID supplied",
            "schema": {
              "$ref": "#/definitions/ApiErrorResponse"
            }
          },
          "404": {
            "description": "demand not found",
            "schema": {
              "$ref": "#/definitions/ApiErrorResponse"
            }
          },
          "405": {
            "description": "Validation exception",
            "schema": {
              "$ref": "#/definitions/ApiErrorResponse"
            }
          }
        },
        "security": [
          {
            "ilv_auth": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "ilv_auth": {
      "type": "basic"
    }
  },
  "definitions": {
...
    "InsertUpdateOffer": {
      "type": "object",
      "properties": {
        "demandId": {
          "type": "integer",
          "format": "int64"
        },
        "startDate": {
          "type": "string",
          "description": "Date when the contract/delivery (case that is signed) begins. Use date format, DD.MM.YYYY"
        },
        "deliveryDate": {
          "type": "string",
          "description": "Date when the contract (case that is signed) expires. Use date format, DD.MM.YYYY"
        },
        "contractType": {
          "type": "string",
          "enum": [
            "SERVICE_CONTRACT",
            "WORK_AND_SERVICE_CONTRACT",
            "SUBSCRIPTION_CONTRACT"
          ]
        },
        "targetAccount": {
          "type": "string",
          "description": "Show name is Zielkontierung"
        },
        "paymentType": {
          "type": "string",
          "enum": [
            "MONTHLY",
            "OBJECTIVE_ORIENTED"
          ]
        },
        "price": {
          "type": "string",
          "description": "100.02 precision of 2 decimals"
        },
        "description": {
          "type": "string"
        },
        "descriptionDocumentUrl": {
          "type": "string"
        }
      }
    },
...
    "ApiErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "ObjectId": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        }
      }
    },
...
  },
...
}